Refactored flowgraph for Oresat0.5
Big refactor of GnuRadio flowgraph for OreSat0.5. We're not receiving EDL or beacons from the older flowgraph, so here's the first cut
at trying to figure out what's going on. We visually simplified the RX and TX flowgraphs, removing hierarchical blocks and lots of
virtual syncs/sources to try to more careful follow the data. Renamed many variables, and moved as many as possible to the config
file. Removed strange "starts_with" check which on RX PDUs which may or may not have been a major mistake. On TX path, replaced
deviation with tx_modulation_index. Calculated should be the same as they were before this. This is probably totally broken and must
be tested before use. There's also some shenanigans around RX and TX sensitivity, which Glenn is going to checkout from the original
textbook source which we forget the name of.